+++ src/main/java/com/example/JobCreation/model/JobVacancies.java
@@ -15,13 +15,13 @@ public class JobVacancies {
 
     private Integer special_cat_id;
     private Integer reservation_cat_id;
-    private Integer location_id;
+    private Long location_id;
     private Integer no_of_vacancies;
 
     public JobVacancies() {
     }
 
-    public JobVacancies(Integer job_vacancy_id, UUID position_id, Integer special_cat_id, Integer reservation_cat_id, Integer location_id, Integer no_of_vacancies) {
+    public JobVacancies(Integer job_vacancy_id, UUID position_id, Integer special_cat_id, Integer reservation_cat_id, Long location_id, Integer no_of_vacancies) {
         this.job_vacancy_id = job_vacancy_id;
         this.position_id = position_id;
         this.special_cat_id = special_cat_id;
@@ -62,11 +62,11 @@ public class JobVacancies {
         this.reservation_cat_id = reservation_cat_id;
     }
 
-    public Integer getLocation_id() {
+    public Long getLocation_id() {
         return location_id;
     }
 
-    public void setLocation_id(Integer location_id) {
+    public void setLocation_id(Long location_id) {
         this.location_id = location_id;
     }
 
@@ -78,4 +78,3 @@ public class JobVacancies {
         this.no_of_vacancies = no_of_vacancies;
     }
 }
-
